+++ lib/more_libs/setting_v2.0.0/thancoder_about_widget.dart
@@ -51,6 +51,22 @@ class ThancoderAboutWidget extends StatelessWidget {
         ),
         Divider(),
         ListTile(
+          leading: Icon(Icons.menu_book),
+          title: Text('Package[pubspec.yaml]'),
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => MarkdownReader(
+                  assetFileName: 'pubspec.yaml',
+                  title: Text('Package[pubspec.yaml]'),
+                ),
+              ),
+            );
+          },
+        ),
+        Divider(),
+        ListTile(
           leading: Icon(Icons.telegram),
           title: Text('Developer'),
           onTap: () {